feat: api 호출 추상화
