+++ src/main/java/shopping/common/client/ProfanityClient.java
@@ -1,0 +1,5 @@
+package shopping.common.client;
+
+public interface ProfanityClient {
+    boolean containsProfanity(String text);
+}
