+++ src/test/java/shopping/product/api/command/ProductCommandControllerTest.java
@@ -13,8 +13,9 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
+import shopping.common.client.ProfanityClient;
 import shopping.product.api.command.dto.ProductRegisterRequest;
-import shopping.product.service.FakeProductNameValidator;
+import shopping.product.service.FakeProfanityClient;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -31,8 +32,8 @@ class ProductCommandControllerTest {
     static class TestConfig {
         @Bean
         @Primary
-        public FakeProductNameValidator fakeProductNameValidator() {
-            return new FakeProductNameValidator();
+        public ProfanityClient fakeProfanityClient() {
+            return new FakeProfanityClient();
         }
     }
 
@@ -43,11 +44,11 @@ class ProductCommandControllerTest {
     private ObjectMapper objectMapper;
 
     @Autowired
-    private FakeProductNameValidator productNameValidator;
+    private FakeProfanityClient profanityClient;
 
     @BeforeEach
     void setUp() {
-        productNameValidator.setProfane(false);
+        profanityClient.setProfane(false);
     }
 
     @Test
@@ -99,7 +100,7 @@ class ProductCommandControllerTest {
     @DisplayName("상품명에 비속어가 포함되면 400을 반환한다")
     void test04() throws Exception {
         // given
-        productNameValidator.setProfane(true);
+        profanityClient.setProfane(true);
         ProductRegisterRequest request = new ProductRegisterRequest("badword", 10000L, "https://example.com/image.jpg");
 
         // when & then
