+++ src/test/java/shopping/product/service/FakeProfanityClient.java
@@ -1,0 +1,16 @@
+package shopping.product.service;
+
+import shopping.common.client.ProfanityClient;
+
+public class FakeProfanityClient implements ProfanityClient {
+    private boolean profane = false;
+
+    public void setProfane(boolean profane) {
+        this.profane = profane;
+    }
+
+    @Override
+    public boolean containsProfanity(String text) {
+        return profane;
+    }
+}
